referral_project/requirements.txt
Django==3.2.3
django-redis==5.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0